Refactor to move `filter_codes` to `CodeList` and remove utils.py
